Work around race condition in tests.
When the file watcher sees the file disappear, it [waits for it to
reappear][1].  But `UpdateEmailFileViaRenameAndReplace` removes the
file entirely after renaming-and-replacing it, and it's possible
that it does that before the file watcher gets around to waiting
for the replacement, so the file watcher hangs forever waiting for
a replacement that will never happen.

As a crude workaround, just wait a long time in the test before
deleting the file, to give the file watcher a high chance of finding
the replacement while it exists.

[1]: https://github.com/sortable/oauth2_proxy/blob/30efbdd6e8788d15f1c960e4f6ac5b4c1b594caa/watcher.go#L55